fix: improve quality of all fractal samplers
Because we are using the same seed for each octave, artifacts are
present in the resulting noise space, represented as diagonal or
axis-aligned grid artifacts, especially noticeable at the domain origin,
and most especially when sourced with grid-based lattice noises, such as
Perlin.

Because fractals are sourced with a sampler instance, we can't easily
change the seed and re-permute each noise sampler. This is because noise
samplers are leaf nodes, and a fractal source could consist of a long
pipeline of modifiers, fractals, and noises combined. It would be
infeasible to traverse the graph of arbitrary types to find which fields
are samplers, and re-permute them.

Instead, we do something much easier. Each octave's domain coordinates
are translated by a random vector that is pre-computed at fractal
sampler construction time. The random vector's components are in the
range [0, 1], as to not waste precision bits in the floating point
values for successive octaves.